+++ tests/test_model_loading.py
@@ -1,6 +1,7 @@
 """Unit tests for the `model_loading` module."""
 
 import pytest
+import torch
 from scandeval.config import ModelConfig
 from scandeval.enums import Framework
 from scandeval.exceptions import InvalidBenchmark, InvalidModel
@@ -80,6 +81,33 @@ def test_load_non_existing_model(dataset_config, benchmark_config):
         languages=[DA],
         model_type="fresh",
         model_cache_dir="cache_dir",
+    )
+    with pytest.raises(InvalidModel):
+        load_model(
+            model_config=model_config,
+            dataset_config=dataset_config,
+            benchmark_config=benchmark_config,
+        )
+
+
+def test_load_awq_model(awq_generative_model_id, dataset_config, benchmark_config):
+    """Test loading an AWQ quantised model."""
+    model_config = get_model_config(
+        model_id=awq_generative_model_id, benchmark_config=benchmark_config
+    )
+    with pytest.raises(InvalidModel):
+        load_model(
+            model_config=model_config,
+            dataset_config=dataset_config,
+            benchmark_config=benchmark_config,
+        )
+
+
+@pytest.mark.skipif(condition=not torch.cuda.is_available(), reason="No GPU available.")
+def test_load_gptq_model(gptq_generative_model_id, dataset_config, benchmark_config):
+    """Test loading a GPTQ quantised model."""
+    model_config = get_model_config(
+        model_id=gptq_generative_model_id, benchmark_config=benchmark_config
     )
     with pytest.raises(InvalidModel):
         load_model(